Trimmed to primary
Changed the structure to be based on primary. Don't wanna do many other changes in semantics or modules until it is runnning properly. Updated values for demand, uniformized the information to Energy Institute 2025. Need data on proper fossil fuel demand (excluding electricity)
Added an allocator to biofuel, same as in fuel extraction